requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.26.2,<0.27
djangorestframework-simplejwt>=5.0.0,<5.1.0